dev-requirements.txt
jinja2==3.0.3
packaging==21.3
pyyaml==6.0
requests==2.27.1
retry==0.9.2
setuptools==59.2.0
twine==3.8.0
vcversioner==2.16.0.0
waiting==1.4.1
wheel==0.37.1
